+++ Source/Krypton Components/Krypton.Toolkit/View Base/Internal/Editors/InternalDesignerEditorButtonBarPanel.Designer.cs
@@ -17,6 +17,7 @@ internal partial class InternalDesignerEditorButtonBarPanel
     {
         kbEdge = new KryptonBorderEdge();
         kcmbTheme = new KryptonThemeComboBox();
+        kbtnThemeSettings = new KryptonButton();
         flpExtraButtons = new FlowLayoutPanel();
         kbtnOk = new KryptonButton();
         kbtnCancel = new KryptonButton();
@@ -34,6 +35,14 @@ internal partial class InternalDesignerEditorButtonBarPanel
         kcmbTheme.DropDownStyle = ComboBoxStyle.DropDownList;
         kcmbTheme.Name = "kcmbDesignerEditorTheme";
         kcmbTheme.TabIndex = 0;
+        //
+        // kbtnThemeSettings
+        //
+        kbtnThemeSettings.Name = "kbtnThemeSettings";
+        kbtnThemeSettings.Size = new Size(90, 28);
+        kbtnThemeSettings.TabIndex = 3;
+        kbtnThemeSettings.Values.Text = "Theme...";
+        kbtnThemeSettings.Visible = false;
         //
         // flpExtraButtons
         //
@@ -65,6 +74,7 @@ internal partial class InternalDesignerEditorButtonBarPanel
         Controls.Add(kbtnCancel);
         Controls.Add(kbtnOk);
         Controls.Add(flpExtraButtons);
+        Controls.Add(kbtnThemeSettings);
         Controls.Add(kcmbTheme);
         Controls.Add(kbEdge);
         Name = "KryptonDesignerEditorButtonBarPanel";
@@ -78,6 +88,7 @@ internal partial class InternalDesignerEditorButtonBarPanel
 
     private KryptonBorderEdge kbEdge;
     private KryptonThemeComboBox kcmbTheme;
+    private KryptonButton kbtnThemeSettings;
     private FlowLayoutPanel flpExtraButtons;
     private KryptonButton kbtnOk;
     private KryptonButton kbtnCancel;